(maint) Remove el_version method
This commit removes the `el_version` method. We only use this in a few
isolated instances in old and unused code. Though this commit doesn't
remove all the old and unused code, it does remove this bit. This is
because this particular piece of code is the only bit in version.rb that
really isn't needed.