(RE-8835) Restructure artifact paths before shipping
This commit adds in the ability to restructure artifact paths before
shipping. We need to be able to allow vanagon and other package building
tools to not care about the paths the artifacts are staged under. To do
this, we need to be able to take a variety of path formats, divine the
corresponding platform tag, and move the package under the new path.

This commit also totally restructures how shipping happens in that it
moves a lot of the actual code into supporting libraries and out of ship
tasks.